// src/app/types.ts
//displayValue
//colocatedType
//aggregate value
export type AggregateValue = "Minimum" | "Maximum" | "Most_Extreme";
export type ColocatedType = "channel" | "aggregate";
export type DisplayValue =
  | "Minimum"
  | "Maximum"
  | "Average"
  | "Median"
  | "5th_Percentile"
  | "95th_Percentile";
export type MetricType = "percent" | "boolean" | "polarity";